requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
backoff==2.2.1
banal==1.0.6
beartype==0.15.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==1.0.1
bokeh==3.6.1
boto3==1.35.57
botocore==1.35.57
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cohere==5.1.8
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
dataset==1.6.2
datasets==3.1.0
decorator==5.1.1
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
executing==2.1.0
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
ftfy==6.3.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-pubsub==2.26.1
google-cloud-run==0.10.10
google-cloud-secret-manager==2.21.0
google-cloud-vision==3.8.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h3==4.1.2
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
instructor==1.2.6
iopath==0.1.10
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-ollama==0.2.0
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.142
language_data==1.2.0
layoutparser==0.3.4
linkify-it-py==2.0.3
litellm==1.42.12
llama-cloud==0.1.4
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-retrievers-pathway==0.2.0
llama-parse==0.5.13
lxml==5.3.0
Mako==1.3.6
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.31.0
msg-parser==1.2.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
Office365-REST-Python-Client==2.5.13
olefile==0.47
ollama==0.3.3
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.0
openai==1.54.3
opencv-python==4.10.0.84
openparse==0.5.6
openpyxl==3.1.5
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
orderly-set==5.2.2
orjson==3.10.11
packaging==23.2
pandas==2.2.3
panel==1.5.3
param==2.1.1
parso==0.8.4
pathway==0.15.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pikepdf==9.4.0
pillow==10.4.0
pillow_heif==0.20.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.6.1
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.13
pypandoc==1.14
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
pysbd==0.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-pptx==0.6.23
python-sat==1.8.dev13
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
ragas==0.2.4
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
sqlglot==10.6.1
srsly==2.4.8
stack-data==0.6.3
striprtf==0.0.26
surya-ocr==0.4.14
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typer==0.13.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unstructured==0.14.3
unstructured-client==0.25.9
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.17.1
zipp==3.20.2
